requirements.txt
torch
transformers
bitsandbytes
auto-gptq
sentence-transformers
gradio
playwright
openai
langchain
langgraph
fastapi
uvicorn
stripe
pydantic
requests
requests_toolbelt
json5
pandas
beautifulsoup4
pyyaml
tomli
Pillow

accelerate
sentencepiece
psutil
faiss-cpu
colorlog
